Correct Cygwin compile error on fnstsw %eax by John4096
git-svn-id: file:///home/jj/hercules.svn/trunk@5293 956126f8-22a0-4046-8f4a-272fa8102e63

+++ ieee-w32.h
@@ -254,7 +254,7 @@ __fpclassifyl (long double x)
 int
 fetestexcept (int excepts)
 {
-  int temp;
+  short temp;
 
 #if defined(NO_INLINE_ASM)
   /* Use CRT function to get current FPU status word */
